ai/requirements.txt
# Gemini API - doğal dil → SQL (Northwind)
google-generativeai>=0.8.0
python-dotenv>=1.0.0
